% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnormimp.R
\name{lnormimp_fit}
\alias{lnormimp_fit}
\title{Fit censored lognormal distribution.
(Uses 'fitdistrplus')}
\usage{
lnormimp_fit(data, censn, cutoff)
}
\arguments{
\item{data}{Numeric data vector}

\item{censn}{Number of lower and upper censored values: c(lower, upper)}

\item{cutoff}{Lower and upper cutoff limits: c(lower, upper)}
}
\value{
'fitdistrplus' distribution fit object
}
\description{
Fit censored lognormal distribution.
(Uses 'fitdistrplus')
}
