% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnormimp.R
\name{lnormimp_sample}
\alias{lnormimp_sample}
\title{Random draws from an interval of a lognormal distribution.
(Uses 'EnvStats')}
\usage{
lnormimp_sample(n, fit, min, max)
}
\arguments{
\item{n}{Number of random draws}

\item{fit}{'fitdistrplus' distribution fit object}

\item{min}{Minimum value for sampling}

\item{max}{Maximum value for sampling}
}
\value{
Vector of sampled values
}
\description{
Random draws from an interval of a lognormal distribution.
(Uses 'EnvStats')
}
